log on register

+++ db/schema.rb
@@ -27,7 +27,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_12_04_164651) do
     t.string "content_type"
     t.text "metadata"
     t.string "service_name", null: false
-    t.bigint "byte_size", null: false
+    t.integer "byte_size", null: false
     t.string "checksum"
     t.datetime "created_at", null: false
     t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
